backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.124.4
greenlet==3.3.0
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
minio==7.2.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9